requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord-webhook==1.1.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.6.0
pycparser==2.21
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0